+++ kernel_development/drivers/reconstructed/zte_tpd/syna_hw_interface_init.c
@@ -1,4 +1,4 @@
-__int64 syna_hw_interface_init(void)
+int syna_hw_interface_init(void)
 {
   int result;
